backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.8.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
google-generativeai==0.8.4